Added files via upload
2nd commit:
Tricker seems to be completed, now it looks for possible dangerous situations favorable to the adversary and try to avoid them, always keeping the preference to build dagerous situations that are favorable to Tricker itself. Tricker is now able to win every other heuristic that we have implemented no matter if it opens the game or not.